+++ packages/zowe-explorer/src/trees/shared/SharedUtils.ts
@@ -13,7 +13,7 @@
 
 import * as vscode from "vscode";
 import * as path from "path";
-import { Gui, IZoweTreeNode, IZoweDatasetTreeNode, IZoweUSSTreeNode, IZoweJobTreeNode, Types, ZosEncoding } from "@zowe/zowe-explorer-api";
+import { Gui, IZoweTreeNode, IZoweDatasetTreeNode, IZoweUSSTreeNode, IZoweJobTreeNode, Types, ZosEncoding, Sorting } from "@zowe/zowe-explorer-api";
 import { UssFSProvider } from "../uss/UssFSProvider";
 import { USSUtils } from "../uss/USSUtils";
 import { Constants } from "../../configuration/Constants";
@@ -21,6 +21,7 @@ import { ZoweLocalStorage } from "../../tools/ZoweLocalStorage";
 import { ZoweLogger } from "../../tools/ZoweLogger";
 import { SharedContext } from "./SharedContext";
 import { Definitions } from "../../configuration/Definitions";
+import { SettingsConfig } from "../../configuration/SettingsConfig";
 
 export class SharedUtils {
     public static async copyExternalLink(this: void, context: vscode.ExtensionContext, node: IZoweTreeNode): Promise<void> {
@@ -230,10 +231,12 @@ export class SharedUtils {
             zosEncoding = await UssFSProvider.instance.fetchEncodingForUri(node.resourceUri);
         }
         let currentEncoding = zosEncoding ? USSUtils.zosEncodingToString(zosEncoding) : await SharedUtils.getCachedEncoding(node);
-        if (zosEncoding?.kind === "binary") {
+        if (zosEncoding?.kind === "binary" || currentEncoding === "binary") {
             currentEncoding = binaryItem.label;
-        } else if (zosEncoding === null || zosEncoding?.kind === "text" || currentEncoding === null || currentEncoding === "text") {
+        } else if (zosEncoding?.kind === "text" || currentEncoding === "text") {
             currentEncoding = ebcdicItem.label;
+        } else if (zosEncoding == null && currentEncoding == null) {
+            currentEncoding = profile.profile?.encoding ?? ebcdicItem.label;
         }
         const encodingHistory = ZoweLocalStorage.getValue<string[]>(Definitions.LocalStorageKey.ENCODING_HISTORY) ?? [];
         if (encodingHistory.length > 0) {
@@ -355,6 +358,62 @@ export class SharedUtils {
                 clearTimeout(timeoutId);
             }
             timeoutId = setTimeout(() => callback(...args), delay);
+        };
+    }
+
+    public static updateSortOptionsWithDefault<T>(sortMethod: T, sortOptions: string[]): void {
+        ZoweLogger.trace("shared.utils.updateSortOptionsWithDefault called.");
+        for (let i = 0; i < sortOptions.length; i++) {
+            sortOptions[i] = sortOptions[i].replace(` ${vscode.l10n.t("(default)")}`, "");
+            if (i === Number(sortMethod)) {
+                sortOptions[i] = `${sortOptions[i]} ${vscode.l10n.t("(default)")}`;
+            }
+        }
+    }
+
+    /**
+     * Gets the sort options from the settings with the default sort option marked
+     * @param sortOptions The list of sort options
+     * @param settingsKey The default sort method key
+     * @param sortMethod The sort method
+     * @returns The list of sort options with the default sort option marked
+     */
+    public static getDefaultSortOptions<T extends object>(sortOptions: string[], settingsKey: string, sortMethod: T): Sorting.NodeSort {
+        ZoweLogger.trace("shared.utils.getDefaultSortOptions called.");
+        const defaultMethod = sortMethod[Object.keys(sortMethod)[0]];
+        const defaultDirection = Sorting.SortDirection.Ascending;
+
+        const sortSetting = SettingsConfig.getDirectValue<Sorting.NodeSort>(settingsKey);
+        if (sortSetting == null || !sortSetting.method || !sortSetting.direction) {
+            SharedUtils.updateSortOptionsWithDefault(defaultMethod, sortOptions);
+            return {
+                method: defaultMethod,
+                direction: defaultDirection,
+            };
+        }
+
+        if (typeof sortSetting.method === "string") {
+            const methodKey = sortSetting.method as keyof typeof sortMethod;
+            if (methodKey in sortMethod) {
+                sortSetting.method = sortMethod[methodKey] as Sorting.JobSortOpts | Sorting.DatasetSortOpts;
+            } else {
+                sortSetting.method = defaultMethod;
+            }
+            SharedUtils.updateSortOptionsWithDefault(sortSetting.method, sortOptions);
+        }
+
+        if (typeof sortSetting.direction === "string") {
+            const directionKey = sortSetting.direction as keyof typeof Sorting.SortDirection;
+            if (directionKey in Sorting.SortDirection) {
+                sortSetting.direction = Sorting.SortDirection[directionKey];
+            } else {
+                sortSetting.direction = defaultDirection;
+            }
+        }
+
+        return {
+            method: sortSetting?.method ?? defaultMethod,
+            direction: sortSetting?.direction ?? defaultDirection,
         };
     }